[refactor] #90 deleteDialogFragment 재사용 가능하게
