+++ app/src/main/res/layout/fragment_feed_delete_dialog.xml
