--- app/src/main/res/layout/fragment_myfeed_highlevel_delete_dialog.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:paddingTop="30dp"
-        android:paddingHorizontal="24dp"
-        android:paddingBottom="18dp"
-        android:elevation="0dp"
-        android:stateListAnimator="@null"
-        android:background="@drawable/shape_gray_fill_12_rect"
-        android:layout_gravity="center">
-
-        <ImageView
-            android:id="@+id/iv_dialog_caution"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:scaleType="centerCrop"
-            android:src="@drawable/ic_wineyfeed_caution"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/tv_dialog_title"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="14dp"
-            android:textColor="@color/gray_900"
-            android:textAppearance="@style/TextAppearance.WINEY.body_b_16"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/iv_dialog_caution"
-            android:text="@string/myfeed_dialog_title" />
-
-        <TextView
-            android:id="@+id/tv_dialog_sub"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="6dp"
-            android:textColor="@color/gray_500"
-            android:textAlignment="center"
-            android:textAppearance="@style/TextAppearance.WINEY.detail_m_13"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_dialog_title"
-            android:text="@string/myfeed_dialog_highlevel_sub" />
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/btn_dialog_cancel"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="20dp"
-            android:background="@drawable/shape_gray_fill_5_rect"
-            android:paddingHorizontal="43dp"
-            android:paddingVertical="10dp"
-            android:textColor="@color/gray_500"
-            android:stateListAnimator="@null"
-            android:text="@string/myfeed_dialog_cancel"
-            android:textAppearance="@style/TextAppearance.WINEY.body_m_16"
-            app:layout_constraintEnd_toStartOf="@id/btn_dialog_delete"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_dialog_sub" />
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/btn_dialog_delete"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="4dp"
-            android:layout_marginTop="20dp"
-            android:background="@drawable/shape_yellow_fill_5_rect"
-            android:paddingHorizontal="29dp"
-            android:paddingVertical="10dp"
-            android:stateListAnimator="@null"
-            android:text="@string/myfeed_dialog_delete"
-            android:textAppearance="@style/TextAppearance.WINEY.body_m_16"
-            android:textColor="@color/gray_900"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@id/btn_dialog_cancel"
-            app:layout_constraintTop_toBottomOf="@id/tv_dialog_sub" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
